Debug
Former-commit-id: 1caa59c4c3aea2bfc9f78c7eeb8799a79b501e41

+++ htdocs/bimpcommercial/objects/ObjectLine.class.php
@@ -6080,7 +6080,7 @@ class ObjectLine extends BimpObject
 
         if ((int) $this->id_product) {
             $product = $this->getProduct();
-            if ($product->getData('type2') == 20) {
+            if (BimpObject::objectLoaded($product) && $product->getData('type2') == 20) {
                 $this->deleteSousLigne($errors, $warnings);
             }
         }
